class inheritance for repos
