+++ backend/src/planner/db/repo/user_repo.py
@@ -1,32 +1,20 @@
 """Module that provides repository for user"""
 
-from planner.db.create_collection import create_collection
+from planner.db.repo.repo import Repo
 from planner.db.schema.user_schema import enforce_user_schema
 
 
-class UserRepo:
+class UserRepo(Repo):
     """Class that provides user repository to interact with the database"""
 
     def __init__(self, db, session):
-        self.db = db
-        self.session = session
-        create_collection(self.db, "users")
+        super().__init__(db, session, db.users, "users")
         enforce_user_schema(self.db)
 
     def find_one_by_email(self, email):
         """Function that finds one user by email"""
 
         return self.db.users.find_one({"email": email}, session=self.session)
-
-    def find_one_by_id(self, _id):
-        """Function that finds one user by email"""
-
-        return self.db.users.find_one({"_id": _id}, session=self.session)
-
-    def insert_one(self, user):
-        """Function that inserts one user"""
-
-        return self.db.users.insert_one(user, session=self.session)
 
     def update_one_by_email(self, email, update):
         """Function that updates one user"""
@@ -36,17 +24,3 @@ class UserRepo:
             update,
             session=self.session,
         )
-
-    def update_one_by_id(self, _id, update):
-        """Function that updates one user"""
-
-        return self.db.users.update_one(
-            {"_id": _id},
-            update,
-            session=self.session,
-        )
-
-    def delete_one_by_id(self, _id):
-        """Function that removes one user"""
-
-        return self.db.users.delete_one({"_id": _id}, session=self.session)
